Clean up Adapter names (enum) Better thread statistics floats changed to doubles some unsigned/signed fussing cleaned up directory listings

+++ module/owlib/src/c/ow_bus.c
@@ -73,7 +73,7 @@ printf("\n");
 //printf("\n");
 //}
         /* send out string */
-        if ( write(devfd,bussend,sendlength) != sendlength ) return -EIO; /* Send the bit */
+        if ( write(devfd,bussend,sendlength) != (int)sendlength ) return -EIO; /* Send the bit */
     }
 //printf("SAG written\n");
     /* get back string -- with timeout and partial read loop */
